+++ dungeon/src/starter/RandomDungeon.java
@@ -3,12 +3,12 @@ package starter;
 import contrib.crafting.Crafting;
 import contrib.entities.EntityFactory;
 import contrib.systems.*;
-import contrib.utils.SoundPlayer;
 import contrib.utils.components.Debugger;
 import core.Entity;
 import core.Game;
 import core.level.utils.LevelSize;
 import core.utils.components.path.SimpleIPath;
+import core.utils.sound.SoundPlayer;
 import java.io.IOException;
 import java.util.logging.Level;
