+++ dungeon/src/starter/RoomBasedDungeon.java
@@ -4,13 +4,13 @@ import contrib.crafting.Crafting;
 import contrib.entities.EntityFactory;
 import contrib.level.generator.graphBased.RoomBasedLevelGenerator;
 import contrib.systems.*;
-import contrib.utils.SoundPlayer;
 import contrib.utils.components.Debugger;
 import core.Entity;
 import core.Game;
 import core.level.elements.ILevel;
 import core.level.utils.DesignLabel;
 import core.utils.components.path.SimpleIPath;
+import core.utils.sound.SoundPlayer;
 import java.io.IOException;
 import java.util.HashSet;
 import java.util.Set;
