+++ dungeon/src/starter/Starter.java
@@ -5,13 +5,13 @@ import contrib.crafting.Crafting;
 import contrib.entities.EntityFactory;
 import contrib.hud.dialogs.OkDialog;
 import contrib.systems.*;
-import contrib.utils.SoundPlayer;
 import core.Entity;
 import core.Game;
 import core.components.PlayerComponent;
 import core.level.elements.ILevel;
 import core.utils.components.MissingComponentException;
 import core.utils.components.path.SimpleIPath;
+import core.utils.sound.SoundPlayer;
 import dsl.interpreter.DSLEntryPointFinder;
 import dsl.interpreter.DSLInterpreter;
 import entrypoint.DSLEntryPoint;
