+++ game/src/core/systems/LevelSystem.java
@@ -1,7 +1,5 @@
 package core.systems;
 
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.audio.Sound;
 import core.Entity;
 import core.Game;
 import core.System;
@@ -19,6 +17,8 @@ import core.utils.components.MissingComponentException;
 import core.utils.components.draw.Painter;
 import core.utils.components.draw.PainterConfig;
 import core.utils.components.path.IPath;
+import core.utils.components.path.SimpleIPath;
+import core.utils.sound.SoundPlayer;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
@@ -66,9 +66,6 @@ public final class LevelSystem extends System {
   private final Painter painter;
   private final Logger levelAPI_logger = Logger.getLogger(this.getClass().getSimpleName());
   private IGenerator generator;
-
-  // Singleton pattern to make junit tests work:
-  private Sound DOOR_SOUND;
 
   /**
    * Create a new {@link LevelSystem}.
@@ -240,12 +237,7 @@ public final class LevelSystem extends System {
   }
 
   private void playSound() {
-    if (DOOR_SOUND == null) {
-      DOOR_SOUND = Gdx.audio.newSound(Gdx.files.internal("sounds/enterDoor.wav"));
-    }
-    long soundId = DOOR_SOUND.play();
-    DOOR_SOUND.setLooping(soundId, false);
-    DOOR_SOUND.setVolume(soundId, 0.3f);
+    SoundPlayer.playSound(new SimpleIPath("sounds/enterDoor.wav"), false, 0.3f);
   }
 
   /**
